move the listing of businesses out to its own component
